vSphere: ignore canceled VMs when scheduling
Now that the migration runner attempts to schedule
as many VMs as it can in a single reconcilation,
it has exposed a bug in the scheduler where canceled
VMs are rescheduled endlessly in certain pathological
cases, preventing the reconciliation from terminating.

(for instance, if a VM is canceled before
it has been marked started, it will always appear
ready to schedule although there is nothing for
the plan controller to do with it.)

The endless rescheduling causes the controller to
infinite loop and be unable to reconcile resources,
necessitating that the offending Migration resource
be deleted and the controller pod be restarted.

Checking for the `Canceled` condition on the VM is adequate
to prevent the problem. (Requesting cancelation of the VM
via the Migration resource will cause the VM to be
marked with the Canceled condition in memory as the Plan
is reconciled, which will cause it to be ignored by the
scheduler, and on the following reconcile it will be
cleaned up as intended.)

Signed-off-by: Sam Lucidi <slucidi@redhat.com>