MTV-1652
Issue:
VM snapshot status in the source provider  isn't being synced upon snapshot deletion.

Fix:
Snapshot casting in the model fails when a nil snapshot object is being updated after deletion.
In that case we need to set the snaphot object in the model to nil  which will be reflected in the provider status

Ref: https://issues.redhat.com/browse/MTV-1652
Signed-off-by: ehazan <ehazan@redhat.com>