MTV-2200 | Add shared disks to guest conversion
Issue: When migrating the shared disks the VM without the shared disk
is missing the conversion step. This causes multiple problem such as
outdated fstab, which causes a system to panic.

Fix: We need to attach the already migrated shared disk to the guest
conversion pod. But as the disk will be converted multiple times and can
be converted multiple times at once we need protect it from multiple
writes at once. This is done by adding a overlay to the image.

Ref: https://issues.redhat.com/browse/MTV-2200

Signed-off-by: Martin Necas <mnecas@redhat.com>