MTV-2346 | Fix fstab with device name config warm migration
Issue: After warm migration of VM which has mounted disk in fstab by the
device name (/dev/sdb) the warm migraiton does not change the fstab
config and ends up in the rescue mode.

Fix: The virt-v2v ignores the fstab disks if the disk in the libvirt
domain has a virtio bus, we need to switch to the sata or scsi. This way
we will force the virt-v2v to change the fstab devices to the virtio
paths (vda, vdb...)

Ref: https://issues.redhat.com/browse/MTV-2346

Signed-off-by: Martin Necas <mnecas@redhat.com>