Prepare for provider-specific migrators
This commit prepares the way for breaking up the migration
runner into provider-specific implementations. Provider
specific migrators will now handle the itineraries and progress
through the pipeline steps, and eventually provide implementations
of the phases. The current pipelines have been moved into a
shared BaseMigrator that the migration runner delegates to during
calls to `execute()`, though the phase implementations currently
remain in the migration runner.

During each call the `execute()` the migration runner will call
`ExecutePhase()` on the provider migrator, which will return
a value indicating whether it provides the phase implementation.
If that value is false, it will fall through to the implementation
in the migration runner.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>