Disable inventory watch for ocp provider
When a http request with an 'X-Watch' header is received by the
inventory controller, it uses the 'watch' feature of the database to
stream events over a websocket to the requesting client. For the local
OCP provider, this prevents a security risk because the objects in the
database are collected by the MTV service account, which has access to
objects that are not necessarily accessible to the requesting client.

This 'watch' feature appears to have been an old feature that was never
actually used, and in fact does not seem to be used by the UI today
either, so there should be no downside to returning 501 Not Implemented
for these requests. If this feature is needed in the future for some
reason, we'll need to implement it using the request's authentication
token so that we don't end up exposing privileged information.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>