ocp container: remove k8s manager
We're no longer watching object changes globally from a k8s manager and
inserting them into a database. So we can strip all of the unused
functionality from the ocp collection implementation and simplify things
significantly. There is not even a need for a database since we no
longer use it as a data source.

This introduces one potential risk: we are now returning nil for the
DB() method and a caller will need to be prepared to handle that fact.
But my analysis of the code is that there are no cases where this can
curretnly happen.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>