MTV-1589 | Add option to delete virt-v2v pod
Issue: When migration completes and user decides to delete VM, resources
such as PVC and PV are kept in terminating state because guest
conversion pod exists. The pod finished, however it is still referencing
the PVC used by VM and thus blocking it's deletion. This happens because
of kubernetes `pvc-protection` finalizer.

Fix: Add option for users to define if they want to delete the virt-v2v
pod after migration succeeds. We do it this way because even
when migration succeeds there is no guarantee that the VM will boot.
Removing the guest conversion pod every time would make it impossible to
troubleshoot this problem using virt-v2v logs.

Ref: https://issues.redhat.com/browse/MTV-1589

Signed-off-by: Stefan Olenocin <solenoci@redhat.com>