Add OCP LiveMigrator
The OCP LiveMigrator is an implementation
of the new provider-specific Migrator interface.
If an OCP plan has its type set to `live`, the
migration runner will delegate execution to
this migrator. The migrator needs to copy
the source VM, blank DVs, and other resources
to the destination cluster and then assist
in setting up the intercluster connection
and finally launching the migration by
creating source and target VMIM resources.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>