virt-v2v: extract udev rules for NetworkManager DHCP networks
Use the stored lease files written by network manager to try to match up
the MAC address and IP to a device name. To do this, we search for a
dhcp lease files that mention the expected address, then pick the newest
one and extract the device name from that file.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>