xcopy-populaotr: allow the populator SA fetch PVs
The populator needs to use the PV spec.CSI.volumeHandle because some
storage vendors need that to identify the LUN and fetch its details. The
PVC name or volumeName is not acceptable.
The volumeHadle is by the CSI spec the identifier in the storage system
so it is expected every provider can use that to uniquely identify
that LUN

Signed-off-by: Roy Golan <rgolan@redhat.com>