MTV-2368 | Migration plan hangs and other migration
Issue:
When guest migration fails during the ImageConversion phase and the resulting PVC remains in a Pending state, the migration plan hangs.
This stalled state prevents any subsequent migration plans from starting, leading to a system-wide block on migrations.

Fix:
When a VM contains only NVMe disks (which are currently unsupported),the loop that builds kDisks is skipped entirely.
As a result, the code assumes kDisks contains at least one disk, which is not guaranteed.
This leads to an exception that halts the current plan and any subsequent ones.
A guard clause should be added before accessing kDisks[bootDisk] to ensure it is not empty.

Also added validation to detect NVMe disk bus types, which are not supported yet.
If detected, the VM is flagged as non-migratable.

Ref: https://issues.redhat.com/browse/MTV-2368
Signed-off-by: ehazan <ehazan@redhat.com>