MTV-2451 | Add virt-v2v-inspector step to warm migratioin
Issue:
When runnning a warm migration the guest conversion, is happening after
we transfer all disks. So if there is problem with the guest conversion
we fail late and users need to retry the disk transferes.

Fix:
To work around this issue we can add virt-v2v-inspector which would run
on already running VMs. The requiremet for that step is that we would
need to create a snapshot so the underlaying disk is not locked by the
runnign VM. Additionally the virt-v2v-inspector needs to allow us to run
on running VMs https://issues.redhat.com/browse/RHEL-88543.

Ref: https://issues.redhat.com/browse/MTV-2451

Signed-off-by: Martin Necas <mnecas@redhat.com>