MTV-2315 | Skip guest conversion
Issue: If there is a guest operating system which is not working with the virt-v2v the whole migration fails. This PR adds a way to work around this issue by skipping the guest conversion parts using the plan.spec.skipGuestConversion parameter. This parameter can true or false, defaults to false.

Fix: Skip the guest conversion, this causes a problem that we are right now using the Virtio buses on everything but if we do not install the drivers on the guest, the VM won't boot. I have defaulted to switch to the sata for disk and e1000e for networking.

Ref: MTV-2315

Signed-off-by: Martin Necas <mnecas@redhat.com>