inventory: Fix user client authentication token
Previously, when loading the client config, we attempted to override
the authentication token by setting config.BearerToken to the user's
authentication token. Unfortunately, it turns out that the default
in-cluster config used BearerTokenFile instead of BearerToken, so
there were two competing token methods set on the client config. To
ensure that the client alway uses the provided user token, clear the
BearerTokenFile variable before setting the user token.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>