xcopy-populator: vmkfstools-wrapper now doesn't require a restart of hostd
Motivation
The installation of the Vib required always a restart of
/etc/init.d/hostd to manifest the changes to the /bin filesystem of the
ESXi. This restart takes long time, and affects the connection to
Vsphere API, and may have some implication we would like to avoid.

Modification
The vib now packs the wrapper and will extract it to /opt/redhat, which is
not the /bin in-memory filesystem, hence uppon installation it will
immediately will be presented and ready to use.

Change
- The tools now is promoted to 0.1.0.
- The vmkfstools-wrapper.py is now located in /opt/redhat/vmkfstools-wrapper.py
- The esxcli wrapper is changed accordingly to point at that new location
- The ansible or any manual installation doesn't require a restart of
  /etc/init.d/hostd daemon anymore.

Signed-off-by: Roy Golan <rgolan@redhat.com>