xcopy-populator: Ensure VIB is installed automatically
Motivation
The VIB is a critical piece of the flow, and it will probably go under
changes as we go. We do have an ansible script to install it, but that
actions is manual. Instead we can try to always look for the vib on the
ESX in context, and check the version, and remediate if necessary. That
way we are always updated and aligned to the populator version.

Modification
With the former fix making the wrapper VIB not needing a restart of
hostd, the installation process becomes quicker and safer.
The populator will fetch the VIB version and compare to the desired
version. If not there it will upload the vib, from the container
filesystem, to the same datastore the VM is on, and then install it,
all through remote ESX commands through the API.

Result
The populator pod automatically checks and installs the VIB on each
attempt to run clone the disk. We pay few seconds in the worst case,
just once per ESX.

Signed-off-by: Roy Golan <rgolan@redhat.com>