xcopy-populator: Add Pure FlashArray provider
This work adds support for Pure FlashArray.

Main difference of this provider to others is that it doesn't allow one
host to belong to more than one hostGroups. Hence, map and unmap will be
performed on the host itself, and not use xcopy-esxs groups. For this
reason the EnsureClonnerIgroup for this provider returns a mapping
context with a list of hosts that match the ESX IQN or WWN. This list is
used later by the map and unmap.

Tested with iSCSI.
Untested with FC.

Signed-off-by: Roy Golan <rgolan@redhat.com>