inventory: Don't return error when a single provider fails
When getting an error for single provider, we were returning an error
response status, which caused the UI to display an "inventory cannot
be reached" error rather than showing the information for the providers
that succeeded. Change this to log the error but still return partial
data to provide a better user experience.

Fixes https://issues.redhat.com/browse/MTV-2701

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>