xcopy-popualtor: For ontap, resolve luns using PV's internalName (#1960)
Motivation
Using the PVs volumeHandle may not be enough for some ontap
installation. The name of the LUN in ontap is set by the CSI driver, and
the default prefix, "trident_pvc_" that is controlled by the trident
configuration can be different. That breaks the resolve logic.

Modification
When using ontap, trident CSI is setting the real name of the lun using
the volume attribute 'internalName'. Using that looks is more reliable
than the heuristic with the volumeHandle

Result
ontap provider will work with whatever trident configuration
`TridentBackend.config.ontap_config.storage_prefix` is set to.

https://issues.redhat.com/browse/ECOPROJECT-2880

Signed-off-by: Roy Golan <rgolan@redhat.com>

Signed-off-by: Roy Golan <rgolan@redhat.com>