MTV-2671 | Migration from RHV to OpenShift Virtualization is failing with error "Running and RunStrategy are mutually exclusive"
Issue:
Creating a virtual machine with "Operating System" type "Linux" in RHV, the inventory cache of forklift have other_linux as OS type.
This will not match any cluster preference and fallback to template selection, For template selection, it will search for templates with label rhel8.1 .
That will be rhel8 templates and in OpenShift Virtualization < 4.18, these templates have "running" in spec which shuld be deprecated but we still need to backword compatible.

Fix:
Adding running to template and identify which methos is being used in the template and update accordingly .

ref: https://issues.redhat.com/browse/MTV-2671
Signed-off-by: Elad <ehazan@redhat.com>