MTV-2707 | xcopy-popualtor: Pass the vmId to the populator
Motivation
Up until now the populator was deducing the name of the VM from the path
of the vmdk file. That approach breaks when 1. a VM is renamed 2. the
disk is a shared disk 3. there are more vm's with same name on another
DC. Probably more cases exists.

Modification
Forkflift will add the vmId to the VsphereXcopyVolumePopulator resource,
which now have a new field, VmId. From there on that is passed all the
way to the command line argument --source-vm-id and there its used to
find the ESX by VM ID.

Result
The migration runs exactly the same, there are no visible changes, or UX
changes, this is all transparent to the user.
Renamed or imported VMs or any of the cases mentioned above will migrate
without any problems.

https://issues.redhat.com/browse/MTV-2707

Signed-off-by: Roy Golan <rgolan@redhat.com>