inventory: reflect runtime errors on inventory provider
Currently, the ocp provider queries data on-demand, which means that it
can fail due to permissions issues or other reasons when somebody issues
a web request to the inventory controller. These errors result in an
http response code sent back to the requestor, but it's not very visible
to the user that there's anything wrong with the provider. To improve
the situation, we store the latest error status for fetching the various
provider resources within the provider's collector. The inventory
controller then checks for these errors as part of the reconcile
process and sets conditions on the provider if there are any current
runtime errors. This required adding a new Collector.GetRuntimeErrors()
interface method, which is currently only implemented by the ocp
provider.