MTV-2715 | OCP provider fails with "Inventory server isn't reachable"
Issue:
When creating OCP provider  it  failes to verify certificate  signed by unknown authority.

Fix:
Because the call to tls.dail()  (which verify internally the server name) has been changed due to a new call with timeout , hence we need to explicitly add the server name to the config to ensure the TLS handshake checks the correct hostname.

Ref: https://issues.redhat.com/browse/MTV-2715
Signed-off-by: Elad <ehazan@redhat.com>