Always use insecure skip verify for ESX hosts (#2073)
Issue:
The line `skipVerifying := base.GetInsecureSkipVerifyFlag(r.Secret)`
tests the host secret for the field `insecureSkipVerify`
We currently don't set this field, and so it is always false.

Fix:
Set the `soap.NewClient` to always use insecure connection.

Future fix:
Document and update UI to set the missing  `insecureSkipVerify` field

Signed-off-by: yaacov <kobi.zamir@gmail.com>