MTV-2855 | Add Dell PowerMax
Add Dell PowerMax provider

The mapping behaviour of this provider is the following:
PowerMax uses the logical container 'MaskingView' to manage hosts and
volumes connections. It does so by defining a StorageGroup, a HostGroup
and a PortGroup.
StorageGroup contains the volumes that needs to be accessed
HostGroup contains the host that needs to access the volumes
PortGroup describe the connection on the storage through which the
connetion will be made, defined by the protocol (mainly)

The provider will ensure that:
A single MaskingView named 'xcopy-esxs' will be created
A single StorqaeGroup named 'xcopy-esxs' will be created and set on the
maskingView
A single HostGroup named 'xcopy-esxs' will be created and set on the
maskingView
A portGroup from the system will be selected by protocol and set on the
masking view

To map a volume add the volume to the storageGroup.
To unmap a volume remove the volume from the storageGroup.

Use a single MaskingView, with a single Storagte
Signed-off-by: Roy Golan <rgolan@redhat.com>