ocp provider: detect connection and permissions problems early
Up to now, the only thing that Collector.Test() did for the ocp provider
was build a client. It didn't attempt to connect to the remote server
or issue any requests, so there was no way to detect until later whether
the credentials were incorrect or lacked sufficient permissions. Add a
basic connection test and try to fetch all of the objects that we need
in order to operate properly.

Previously an ocp provider without sufficient permissions would be
stuck in 'staging' phase or would get to 'ready' phase but would produce
errors when fetching the inventory. Now we can detect this earlier and
provide feedback to the user about what is wrong.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>