Make Kubevirt cert sync happen once per plan exec
Add `Begin()` method to the Migration interface that
is called once at the beginning of a plan execution.
This can be used for migration setup tasks that need
to be done once before any of the VMs begin to execute.

Move call to SynchronizeCertificateBundles() into the
OCP LiveMigrator's `Begin()` method.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>