xcopy-populator: use PersistentVolume type instead of volumeHandle
Each provider is using a different part of the k8s PersistentVolume to
identity the underlying volume in their storage system.
Spec.Csi.VolumeHandle is not adopted as such - some vendor use
VolumeAttributes, with different attribute names. Some won't give the
full name, and just use the PV name, with or witout hidden prefix that is known
only to the CSI driver. NetApp trident's will let you see the prefix, pure for
example uses the "px" and their portworx cluster uuid, without any
visibility (other than just reading their csi code).

Passing a slimmed down PersistentVolume object leaves the resolution of
the volume in the hands of the provider code, which is safer to
change for one without affecting the others.

Signed-off-by: Roy Golan <rgolan@redhat.com>