Improve error reporting for ocp providers (#2097)
When creating a new ocp provider, if it is not configured correctly, it
would
often get stuck in 'staging' state and never give the user any
indication of
why wasn't working. This patch series attempts to improve the situation
and
make the problem more debuggable to a user.

The first patch implements an actual connection test for ocp providers
which
didnt' exist before and should allow the user to catch a lot of
configuration
errors early.

The second patch may not be necessary since the first patch may be
sufficient.
But it keeps a map of errors that were encountered during inventory
requests and
provides an interface for the provider reconciler to get that
information and
use it to report Conditions on the provider object.

Fixes: https://issues.redhat.com/browse/MTV-2447

---------

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>