Controller: fetch domain xml from libvirt
Rather than constructing our own libvirt xml to pass to virt-v2v, we
should be using libvirt itself to connect to the vsphere server and
generate a domain xml file for the domain that we are trying to migrate.
This ensures that virt-v2v has as complete information as possible to
perform conversion of the domain.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>