controller: Add DomainXML() method to Builder interface
This allows providers to provide a method for getting libvirt domain
xml describing a specific vm. Currently will  only be implemented
for vmware using the libvirt vmware driver. The libvirt xml is used
by virt-v2v to convert the vm and it's important to have as accurate
information about the vm during conversion as possible.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>