Wait for target VMI to be ready to avoid race (#2200)
An admission webhook will reject the target
VirtualMachineInstanceMigration unless the target VM's VMI has already
been created and is waiting in the WaitForSync state. To avoid this,
explictly wait for the target VMI to be ready before creating the VMIM.

This wasn't being caught before because using a 3rd cluster to
coordinate resulted in enough latency that the VMI was present before
the next migration step took place.

This also adds another role that is necessary for the host cluster to be
able to function as the source or target of the migration.

---------

Signed-off-by: Sam Lucidi <slucidi@redhat.com>