plan validator: use unused conditions
Both `powerStateUnsupported` and `vmMigrationTypeUnsupported` variables
were written to (i.e. the 'Items' member was set), but the struct itself
was never used. All other similar condition structs were used to set
conditions on the plan status when the Items was not empty, so it seems
that this just got forgotten.

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>