build: Remove LIBGUESTFS_BACKEND=direct everywhere
This should *never* be used.  It causes libguestfs to use an untested
code path.  It also stops libvirt from containerizing the qemu
subprocess, reducing security.

Wherever this is seen, it should be removed.  If doing so causes a
problem (which is unlikely as it puts us back on the tested code path)
then file a detailed bug about the problem, but do not reintroduce the
environment variable.

Signed-off-by: Richard W.M. Jones <rjones@redhat.com>