MTV-3035 | Switch MTV LIBGUESTFS_BACKEND to libvirt
Issue:
The LIBGUESTFS_BACKEND=direct is not tested and not supported path, we
need to switch to LIBGUESTFS_BACKEND=libvirt:qemu:///session.

This switch caused multiple issues:
- https://issues.redhat.com/browse/RHEL-105490
- https://issues.redhat.com/browse/RHEL-105529

Both of them are caused by running the libguestfs inside container.

For RHEL-105490 we need to set home directly, becuase QEMU users as
which this container is running does not have permissions to write to
the root directory. We will create home directory for that user.

For RHEL-105529 the virt-v2v had problem cleaning up the libguestfs VMs
in the contianer. This is cause as there is not init process PID 1 which
would cleanup the zombie processes. For this we need to use some init
process such as `catatonit` and run it before the virt-v2v.

Signed-off-by: Martin Necas <mnecas@redhat.com>