MTV-3078 | Power off VM fails due to "VMware Tools is not running"
Issue:
Cold migration of a VM with LUKS-encrypted disks and skipGuestConversion: true fails because VMware Tools is not running. VMware Tools cannot start until the encrypted disk is unlocked, but with skipGuestConversion: true, the user typically doesn't have the decryption password and it remains locked.

Fix:
Adding a fall-back to performing a hard power off. This ensures the migration can proceed even if graceful shutdown isn't possible.

Ref: https://issues.redhat.com/browse/MTV-3078
Signed-off-by: Elad Hazan <ehazan@redhat.com>