MTV-2855 | Add Dell PowerMax copy offload provider (#2174)
Add Dell PowerMax provider
    
The mapping behaviour of this provider is the following:
PowerMax uses the logical container 'MaskingView' to manage hosts and
volumes connections. It does so by defining a StorageGroup, a HostGroup
and a PortGroup.
StorageGroup contains the volumes that needs to be accessed
HostGroup contains the host that needs to access the volumes
PortGroup describe the connection on the storage through which the
connetion will be made, defined by the protocol (mainly)
    
The provider will ensure that:
A single MaskingView named 'xcopy-esxs' will be created
A single StorqaeGroup named 'xcopy-esxs' will be created and set on the
maskingView
A single HostGroup named 'xcopy-esxs' will be created and set on the
maskingView
A portGroup from the system will be selected by protocol and set on the
masking view
    
To map a volume add the volume to the storageGroup.
To unmap a volume remove the volume from the storageGroup.

https://issues.redhat.com/browse/MTV-2855


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **New Features**
* Added support for Dell PowerMax storage systems in the vSphere Xcopy
Volume Populator, enabling integration and migration workflows with
PowerMax as a storage backend.
* Extended configuration and CRD schemas to recognize "powermax" as a
valid storage vendor option.

* **Documentation**
* Updated documentation with details on configuring secrets for PowerMax
storage and added troubleshooting guidance for vSphere/ESXi
environments.

* **Bug Fixes**
* Improved file path handling for VIB uploads to enhance compatibility.

* **Chores**
* Updated and added dependencies and supporting files for PowerMax
integration.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Signed-off-by: Roy Golan <rgolan@redhat.com>