MTV-2312 | Maintain network during VM migration with DHCP (#1706)
When migrating a VM with a network interface that uses DHCP, make a best
effort to maintain network connectivity after migration. We use the same
scripts that we use for static IP interface mapping, but attempt to use
dhcp lease files on the vm to determine the interface name associated
with a particular IP and create a udev rule for that device. Note that
this only works for vms that are using NetworkManager or dhclient to
handle DHCP. If using systemd-networkd, the dhcp lease files are not
available in the filesystem while the vm is not running so we don't
handle that scenario.

---------

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>