Fix hanging provider caused by runtime error check
PR #2097 introduced a check for inventory runtime errors during
the validation step in the Provider controller reconciliation
loop. Runtime errors are accumulated when the inventory API endpoints
fail, and the intention is to surface those errors in the Provider CR
status. Unfortunately, this does not work as intended.

The connection test for OCP providers checks that it is possible to access
the resources needed by the inventory API. Because the check for runtime
errors occurs after the connection test, only two outcomes are possible:

1) The connection test fails, in which case any accumulated runtime
errors will not be exposed because the validation process aborts before
that point. In this case the failed connection test condition will report
the same sort of errors and no information has been lost.

2) The connection test succeeds, implying that the credentials or other
issue has been solved. In this case the validation process will move to
the next step and any accumulated runtime errors will now cause validation
to fail. However, the successful connection test almost certainly means
that any runtime errors that accumulated since the last reconcile are
out-of-date and irrelevant. This causes the following error condition
for OCP providers:

1. Token expires
2. Runtime errors accumulate as inventory accesses fail
3. Secret is updated with a new token, triggering reconcile
4. Connection test passes, but runtime errors that accumulated
   previously result in a blocker condition.
5. Blocker condition prevents the inventory collector from being
   updated.
6. Provider remains trapped in an error state despite the token
   being valid and controller logs reporting that the connection
   test is succeeding.

All that is necessary to resolve the problem is to remove the runtime
error check in the validator. As mentioned above, the check is redundant
with the connection test so nothing is lost. In order to accomplish
the goal that the runtime error check is intended to accomplish, a future
improvement should either regularly perform connection tests or else
enqueue an event for the provider when a runtime error occurs so that
a reconcile is triggered.

Signed-off-by: Sam Lucidi <slucidi@redhat.com>